numocity-fullstack/backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
